feat: 사용자 등록 요청을 처리하는 컨트롤러 추가
